+++ DbConfigurator.UI.UnitTests/Features/DistributionInformation/DistributionInformationTableViewModelTests.cs
@@ -4,7 +4,7 @@ using DbConfigurator.Model.Entities.Wrapper;
 using DbConfigurator.UI.Event;
 using DbConfigurator.UI.Features.Areas.Event;
 using DbConfigurator.UI.Features.DistributionInformations;
-using DbConfigurator.UI.Services.Interfaces;
+using DbConfigurator.UI.Services;
 using DbConfigurator.UI.Startup;
 using DbConfigurator.UI.UnitTests.Base;
 using DbConfigurator.UI.ViewModel.Base;
