+++ DbConfigurator.UI/Features/BuisnessUnits/Table/BuisnessUnitTableViewModel.cs
@@ -2,6 +2,7 @@
 using DbConfigurator.Model.DTOs.Wrapper;
 using DbConfigurator.UI.Event;
 using DbConfigurator.UI.Features.Areas.Event;
+using DbConfigurator.UI.Services;
 using DbConfigurator.UI.Services.Interfaces;
 using DbConfigurator.UI.Startup;
 using DbConfigurator.UI.ViewModel.Base;
