+++ DbConfigurator.UI/Features/DistributionInformations/Detail/DistributionInformationDetailViewModel.cs
@@ -2,7 +2,7 @@
 using DbConfigurator.UI.Event;
 using DbConfigurator.UI.Extensions;
 using DbConfigurator.UI.Features.Areas.Event;
-using DbConfigurator.UI.Services.Interfaces;
+using DbConfigurator.UI.Services;
 using DbConfigurator.UI.ViewModel.Base;
 using Microsoft.EntityFrameworkCore.ChangeTracking.Internal;
 using Prism.Commands;
