+++ DbConfigurator.UI/Services/BuisnessUnitService.cs
@@ -1,10 +1,8 @@
 ﻿using DbConfigurator.Model.DTOs.Core;
 using DbConfigurator.Model.Entities.Core;
-using DbConfigurator.UI.Services;
-using DbConfigurator.UI.Services.Interfaces;
 using DbConfigurator.UI.Startup;
 
-namespace DbConfigurator.UI.Features.BusinessUnits
+namespace DbConfigurator.UI.Services
 {
     public class BusinessUnitService : GenericDataService<BusinessUnit, BusinessUnitDto>, IBusinessUnitService
     {
