+++ DbConfigurator.UI/Services/CountryService.cs
@@ -1,15 +1,13 @@
 ﻿using DbConfigurator.Model.DTOs.Core;
 using DbConfigurator.Model.Entities.Core;
-using DbConfigurator.UI.Services;
-using DbConfigurator.UI.Services.Interfaces;
 using DbConfigurator.UI.Startup;
 
-namespace DbConfigurator.UI.Features.Countries
+namespace DbConfigurator.UI.Services
 {
     public class CountryService : GenericDataService<Country, CountryDto>, ICountryService
     {
         public CountryService(AutoMapperConfig autoMapper)
-            //: base(repository, autoMapper)
+        //: base(repository, autoMapper)
         {
         }
     }
