+++ DbConfigurator.UI/Services/DistributionInformationService.cs
@@ -1,7 +1,5 @@
 ﻿using DbConfigurator.Model.DTOs.Core;
 using DbConfigurator.Model.Entities.Core;
-using DbConfigurator.UI.Services;
-using DbConfigurator.UI.Services.Interfaces;
 using DbConfigurator.UI.Startup;
 using Microsoft.EntityFrameworkCore;
 using System;
@@ -9,7 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace DbConfigurator.UI.Features.DistributionInformations
+namespace DbConfigurator.UI.Services
 {
     public class DistributionInformationService : GenericDataService<DistributionInformation, DistributionInformationDto>, IDistributionInformationService
     {
