+++ DbConfigurator.UI/Services/IDistributionInformationService.cs
@@ -1,8 +1,9 @@
 ﻿using DbConfigurator.Model.DTOs.Core;
+using DbConfigurator.UI.Services.Interfaces;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace DbConfigurator.UI.Services.Interfaces
+namespace DbConfigurator.UI.Services
 {
     public interface IDistributionInformationService : IDataService<DistributionInformationDto>
     {
