+++ DbConfigurator.UI/Services/IRecipientService.cs
@@ -1,6 +1,7 @@
 ﻿using DbConfigurator.Model.DTOs.Core;
+using DbConfigurator.UI.Services.Interfaces;
 
-namespace DbConfigurator.UI.Services.Interfaces
+namespace DbConfigurator.UI.Services
 {
     public interface IRecipientService : IDataService<RecipientDto>
     {
