+++ DbConfigurator.UI/Services/RecipientService.cs
@@ -1,15 +1,13 @@
 ﻿using DbConfigurator.Model.DTOs.Core;
 using DbConfigurator.Model.Entities.Core;
-using DbConfigurator.UI.Services;
-using DbConfigurator.UI.Services.Interfaces;
 using DbConfigurator.UI.Startup;
 
-namespace DbConfigurator.UI.Features.Recipients
+namespace DbConfigurator.UI.Services
 {
     public class RecipientService : GenericDataService<Recipient, RecipientDto>, IRecipientService
     {
-        public RecipientService(AutoMapperConfig autoMapper) 
-            //: base(repository, autoMapper)
+        public RecipientService(AutoMapperConfig autoMapper)
+        //: base(repository, autoMapper)
         {
         }
     }
